Update: Changed datatypes to smaller float16 or int32

+++ htmresearch/algorithms/apical_tiebreak_bayesian_summing_temporal_memory.py
@@ -117,12 +117,12 @@ class SummingBayesianApicalTiebreakPairMemory(ApicalTiebreakBayesianTemporalMemo
             seed=seed
         )
 
-        self.basalConnectionCount = np.zeros((1, self.numberOfCells(), self.basalInputSize), dtype='int64')
-        self.apicalConnectionCount = np.zeros((1, self.numberOfCells(), self.apicalInputSize), dtype='int64')
-        self.basalSegmentActivationCount = np.zeros((1, self.numberOfCells()), dtype='int64')
-        self.apicalSegmentActivationCount = np.zeros((1, self.numberOfCells()), dtype='int64')
-        self.basalInputCount = np.zeros(self.basalInputSize, dtype='int64')
-        self.apicalInputCount = np.zeros(self.apicalInputSize, dtype='int64')
+        self.basalConnectionCount = np.zeros((1, self.numberOfCells(), self.basalInputSize), dtype='int32')
+        self.apicalConnectionCount = np.zeros((1, self.numberOfCells(), self.apicalInputSize), dtype='int32')
+        self.basalSegmentActivationCount = np.zeros((1, self.numberOfCells()), dtype='int32')
+        self.apicalSegmentActivationCount = np.zeros((1, self.numberOfCells()), dtype='int32')
+        self.basalInputCount = np.zeros(self.basalInputSize, dtype='int32')
+        self.apicalInputCount = np.zeros(self.apicalInputSize, dtype='int32')
         self.updateCounter = 0
 
     def _addNewSegments(self, isBasal=True):
@@ -225,13 +225,13 @@ class SummingBayesianApicalTiebreakPairMemory(ApicalTiebreakBayesianTemporalMemo
         connection_matrix = np.outer(segments, inputValues)
         # Consider only active segments
         connection_matrix = connection_matrix.reshape(numSegments, self.numberOfCells(), connectionCount.shape[-1])
-        connectionCount += connection_matrix.astype('int64')
+        connectionCount += connection_matrix.astype('int32')
 
         # Updating moving average bias of each segment
-        segmentActivityCount += segments.astype('int64')
+        segmentActivityCount += segments.astype('int32')
 
         # Updating moving average input activity
-        inputCount += inputValues.astype('int64')
+        inputCount += inputValues.astype('int32')
 
         if isBasal:
             self._setBasalConnectionData(
@@ -254,10 +254,10 @@ class SummingBayesianApicalTiebreakPairMemory(ApicalTiebreakBayesianTemporalMemo
         activationCount = self.basalSegmentActivationCount if isBasal else self.apicalSegmentActivationCount
         inputCount = self.basalInputCount if isBasal else self.apicalInputCount
 
-        weights = (connectionCount.astype('float64') * self.updateCounter) / np.outer(
+        weights = (connectionCount.astype('float16') * self.updateCounter) / np.outer(
             activationCount,
             inputCount
-        ).reshape(connectionCount.shape).astype('float64')
+        ).reshape(connectionCount.shape).astype('float16')
         weights[weights == 0] = 1. / float(self.updateCounter)
         weights[np.isnan(weights)] = 1.
         return weights
